+++ CHANGELOG.md
@@ -6,6 +6,7 @@ IMPROVEMENTS:
  * build: Updated to Go 1.17 [[GH-545](https://github.com/hashicorp/nomad-autoscaler/pull/545)]
  * plugins/apm/datadog: Add support for custom server site [[GH-548](https://github.com/hashicorp/nomad-autoscaler/pull/548)]
  * plugins/apm/prometheus: Add support for basic auth and custom headers [[GH-522](https://github.com/hashicorp/nomad-autoscaler/pull/522)]
+ * plugins/apm/prometheus: Add support for TLS CA certificates [[GH-547](https://github.com/hashicorp/nomad-autoscaler/pull/547)]
  * plugins/target/nomad: Reduce log level for active deployments error messages [[GH-542](https://github.com/hashicorp/nomad-autoscaler/pull/542)]
  * policy: Prevent scaling cluster to zero when using the Nomad APM [[GH-534](https://github.com/hashicorp/nomad-autoscaler/pull/534)]
  * scaleutils: Add combined filter to allow filtering by node class and datacenter [[GH-535](https://github.com/hashicorp/nomad-autoscaler/pull/535)]